Worked on splitting flutter stuff into fast_rx_flutter
