+++ fast_rx/fast_rx/lib/fast_rx.dart
@@ -1,7 +1,6 @@
 library fast_rx;
 
 export 'src/exceptions.dart';
-export 'src/fast_builder.dart';
 
 // rx/composite
 export 'src/rx/composite/rx_composite_list.dart';
@@ -15,4 +14,5 @@ export 'src/rx/object/rx_map.dart';
 export 'src/rx/object/rx_set.dart';
 
 export 'src/rx/rx_typedef.dart';
+export 'src/rx/rx_zone.dart';
 export 'src/rx/rx.dart';
