+++ fast_rx/fast_rx_flutter/analysis_options.yaml
@@ -1,4 +1,1 @@
-include: package:flutter_lints/flutter.yaml
-
-# Additional information about this file can be found at
-# https://dart.dev/guides/language/analysis-options
+import: packages:rexios_lints/flutter/package.yaml
