+++ fast_rx/fast_rx_flutter/test/fast_builder_test.dart
@@ -1,4 +1,5 @@
 import 'package:fast_rx/fast_rx.dart';
+import 'package:fast_rx_flutter/fast_rx_flutter.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
