--- fast_rx/fast_rx_flutter/test/fast_rx_flutter_test.dart
@@ -1,12 +1,0 @@
-import 'package:flutter_test/flutter_test.dart';
-
-import 'package:fast_rx_flutter/fast_rx_flutter.dart';
-
-void main() {
-  test('adds one to input values', () {
-    final calculator = Calculator();
-    expect(calculator.addOne(2), 3);
-    expect(calculator.addOne(-7), -6);
-    expect(calculator.addOne(0), 1);
-  });
-}
