+++ fast_rx/fast_rx_test/pubspec.yaml
@@ -5,12 +5,8 @@ homepage: https://github.com/Rexios80/fast_ui/tree/master/fast_rx_test
 
 environment:
   sdk: ">=2.17.0 <3.0.0"
-  flutter: ">=1.17.0"
 
 dependencies:
-  flutter:
-    sdk: flutter
-
   fast_rx: ^0.3.0
   test: ^1.19.0
