+++ fast_ui/example/pubspec.yaml
@@ -34,6 +34,8 @@ dependency_overrides:
     path: ../../fast_overlays
   fast_rx:
     path: ../../fast_rx/fast_rx
+  fast_rx_flutter:
+    path: ../../fast_rx/fast_rx_flutter
   fast_rx_persistence:
     path: ../../fast_rx/fast_rx_persistence
