+++ fast_ui/lib/fast_ui.dart
@@ -4,5 +4,5 @@ library fast_ui;
 export 'package:fast_extensions/fast_extensions.dart';
 export 'package:fast_nav/fast_nav.dart';
 export 'package:fast_overlays/fast_overlays.dart';
-export 'package:fast_rx/fast_rx.dart';
+export 'package:fast_rx_flutter/fast_rx_flutter.dart';
 export 'package:fast_rx_persistence/fast_rx_persistence.dart';
